backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
librouteros==3.2.1
pyyaml==6.0.2
pydantic==2.10.4
pydantic-settings==2.7.1
apscheduler==3.10.4
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
